requirements.txt
torch~=1.10.0
PyYAML~=6.0
numpy~=1.21.3
matplotlib~=3.5.0
tqdm~=4.62.3
scikit-learn~=1.0.1
sklearn~=0.0